PyStrInterned
