+++ vm/src/builtins/mod.rs
@@ -59,7 +59,7 @@ pub(crate) mod bool_;
 pub use bool_::PyBool;
 #[path = "str.rs"]
 pub(crate) mod pystr;
-pub use pystr::{PyStr, PyStrRef};
+pub use pystr::{PyStr, PyStrInterned, PyStrRef};
 #[path = "super.rs"]
 pub(crate) mod super_;
 pub use super_::PySuper;
