+++ vm/src/stdlib/sys.rs
@@ -512,8 +512,8 @@ mod sys {
     }
 
     #[pyfunction]
-    fn intern(s: PyRefExact<PyStr>, vm: &VirtualMachine) -> PyRefExact<PyStr> {
-        vm.ctx.intern_str(s).to_exact()
+    fn intern(s: PyRefExact<PyStr>, vm: &VirtualMachine) -> PyRef<PyStr> {
+        vm.ctx.intern_str(s).to_owned()
     }
 
     #[pyattr]
